+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.0.9] - Not yet released
+
+### Added
+
+- Added highlighting of log() function calls in scripts
+- Added more options to script functions mkDir, rmDir und deleteFile
+
 ## [0.0.8] - 2024-11-06
 
 ### Changed